# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/reservations/hold": {
      "post": {
        "description": "픽업 시각 검증 → 전체 아이템 재고 확인 → All-or-Nothing 방식으로 Redis Hold 저장 (TTL 2분).\n\n**All-or-Nothing 규칙**\n- 요청한 모든 아이템의 재고가 충분할 때만 holdToken을 발급하고 Redis에 Hold를 생성합니다.\n- 단 하나라도 재고 부족이면 Redis Hold를 생성하지 않고 `success: false`와 상세 실패 목록을 반환합니다.\n\n**세션 상태 전이**\n- 성공: `PRE_HOLD_CONFIRM → WAITING_FOR_CONFIRM` (hold_token 함께 저장)\n- 실패: 현재 상태 → `FAIL` (last_error에 실패 사유 기록)\n\n실제 재고 차감은 `/confirm` 단계에서만 수행됩니다.",
        "operationId": "holdReservation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "일부 또는 전체 재고 부족 — holdToken 미발급 (success: false)",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "success": false,
                      "holdToken": null,
                      "items": [
                        {
                          "id": "101",
                          "name": "소금빵",
                          "requestedCount": 2,
                          "heldCount": 2,
                          "status": "SUCCESS"
                        },
                        {
                          "id": "202",
                          "name": "고구마빵",
                          "requestedCount": 3,
                          "heldCount": 0,
                          "status": "OUT_OF_STOCK",
                          "reason": "재고 부족 (남은 수량: 1개)"
                        }
                      ]
                    },
                    "message": "Hold created successfully"
                  }
                }
              }
            }
          },
          "201": {
            "description": "전체 hold 성공 — holdToken 발급",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "data": {
                      "success": true,
                      "holdToken": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "items": [
                        {
                          "id": "101",
                          "name": "소금빵",
                          "requestedCount": 2,
                          "heldCount": 2,
                          "status": "SUCCESS"
                        },
                        {
                          "id": "202",
                          "name": "크림빵",
                          "requestedCount": 1,
                          "heldCount": 1,
                          "status": "SUCCESS"
                        }
                      ]
                    },
                    "message": "Hold created successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "잘못된 픽업 시각 또는 영업시간 외",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Pickup time must be a valid future datetime"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자 또는 매장 없음",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    },
                    "message": {
                      "type": "string",
                      "example": "Store not found"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "재고 임시 hold 생성 (All-or-Nothing)",
        "tags": [
          "Reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bread API",
    "description": "Bread Reservation & Inventory API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateHoldDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "예약 사용자 ID"
          },
          "storeId": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "예약 대상 매장 ID"
          },
          "pickupTime": {
            "type": "string",
            "description": "픽업 예정 시각 (ISO 8601). 현재 시각 이후여야 하며 매장 영업시간(open_time ~ close_time) 범위 내여야 합니다."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "breadId": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "description": "빵 ID"
                },
                "qty": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991,
                  "description": "예약 수량 (1 이상)"
                }
              },
              "required": [
                "breadId",
                "qty"
              ]
            },
            "description": "예약할 빵 목록 (한 매장에서 여러 종류 동시 예약 가능)"
          }
        },
        "required": [
          "userId",
          "storeId",
          "pickupTime",
          "items"
        ]
      },
...
    }
  }
}